extract refinement into partis.refine module
